+++ saleor/graphql/account/mutations/authentication/request_password_reset.py
@@ -1,12 +1,15 @@
+from urllib.parse import urlencode
+
 import graphene
 from django.conf import settings
+from django.contrib.auth.tokens import default_token_generator
 from django.core.exceptions import ValidationError
 from django.utils import timezone
 
 from .....account.error_codes import AccountErrorCode
 from .....account.notifications import send_password_reset_notification
 from .....account.utils import retrieve_user_by_email
-from .....core.utils.url import validate_storefront_url
+from .....core.utils.url import prepare_url, validate_storefront_url
 from .....webhook.event_types import WebhookEventAsyncType
 from ....channel.utils import clean_channel, validate_channel
 from ....core import ResolveInfo
@@ -47,7 +50,11 @@ class RequestPasswordReset(BaseMutation):
             WebhookEventInfo(
                 type=WebhookEventAsyncType.NOTIFY_USER,
                 description="A notification for password reset.",
-            )
+            ),
+            WebhookEventInfo(
+                type=WebhookEventAsyncType.ACCOUNT_SET_PASSWORD_REQUESTED,
+                description="Setting a new password for the account is requested.",
+            ),
         ]
 
     @classmethod
@@ -99,10 +106,21 @@ class RequestPasswordReset(BaseMutation):
     def perform_mutation(cls, _root, info: ResolveInfo, /, **data):
         email = data["email"]
         redirect_url = data["redirect_url"]
-        channel_slug = data.get("channel")
         user = cls.clean_user(email, redirect_url, info)
+        cls.post_save_action(info, user, data)
+        user.last_password_reset_request = timezone.now()
+        user.save(update_fields=["last_password_reset_request"])
 
-        if not user.is_staff:
+        return RequestPasswordReset()
+
+    @classmethod
+    def post_save_action(cls, info: ResolveInfo, instance, cleaned_input):
+        channel_slug = cleaned_input.get("channel")
+        token = default_token_generator.make_token(instance)
+        params = urlencode({"email": instance.email, "token": token})
+        redirect_url = cleaned_input["redirect_url"]
+
+        if not instance.is_staff:
             channel_slug = clean_channel(
                 channel_slug, error_class=AccountErrorCode
             ).slug
@@ -110,14 +128,19 @@ class RequestPasswordReset(BaseMutation):
             channel_slug = validate_channel(
                 channel_slug, error_class=AccountErrorCode
             ).slug
+
         manager = get_plugin_manager_promise(info.context).get()
         send_password_reset_notification(
             redirect_url,
-            user,
+            instance,
             manager,
             channel_slug=channel_slug,
-            staff=user.is_staff,
+            staff=instance.is_staff,
         )
-        user.last_password_reset_request = timezone.now()
-        user.save(update_fields=["last_password_reset_request"])
-        return RequestPasswordReset()
+        cls.call_event(
+            manager.account_set_password_requested,
+            instance,
+            channel_slug,
+            token,
+            prepare_url(params, redirect_url),
+        )