+++ CHANGELOG.md
@@ -97,6 +97,8 @@ Shipping methods can be removed by the user after it has been assigned to a chec
   - Called when user confirm an account with `confirmAccount` mutation.
 - Add `ACCOUNT_DELETED` webhook - #13471, by @Air-t
   - Called after account deletion is confirmed with `accountDelete` mutation.
+- Add `ACCOUNT_SET_PASSWORD_REQUESTED` webhook - #13486, by @Air-t
+  - Called after `requestPasswordReset` or `customerCreate` mutation.
 
 ### Other changes
 - Add possibility to log without confirming email - #13059 by @kadewu